Moved CronAdjuster to common, so it can be used for the rule simulation
