+++ bundles/org.openhab.core/src/test/java/org/openhab/core/internal/scheduler/CronAdjusterMiscTest.java
@@ -17,6 +17,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
+import org.openhab.core.common.CronAdjuster;
 
 /**
  * Test cases for {@link CronAdjuster} to test error handling and misc cases.
